cache_hash: hash function changes
- Replace the explicit Blake3 type with a more generic Hasher alias, so
that the hash function can easily be updated.
- Switch to Poly1305.

Poly1305 hashes are only 128 bit (-> smaller cache), but has
information-theoretic guarantees that practical collisions cannot happen
(differential probability is (L/2)/(2^106) with L being the maximum length
of a file).

Our Poly1305 implementation is currently about 3 times faster than Blake3,
and can be made 6-8 times faster with a vectorized implementation.

We can later replace it with an optimized BLAKE3 implementation if it
is on par from a performance perspective.

The flipside of Poly1305 is that it requires a secret key. If an adversary
can read the cache, collisions become trivial to craft.

So, a key is now created along with a new cache directory.