Implement std.meta.tagPayload
This commit implements std.meta.tagPayload, a utility function which
performs field access on a tagged union if and only if the requested
field is the active field. In the case that it is not, the function
will return null.

We already have TagPayloadType to retrieve the type of a tagged union
field based on its tag, and this is a natural extension of that idea.
The current alternative is essentially the body of the function, which
does not synergize well with providing a default value.

This commit additionally changes the body of TagPayloadType to use
std.meta.fieldInfo function rather than iterating over fields manually,
and adds the comptime keyword to the tag parameter for clarity.