Implement TagPayloadPtrType and tagPayloadPtr
These are alternatives to TagPayloadType and tagPayload, respectively,
instead taking a pointer to a tagged union and returning a derived
pointer corresponding to the enum tag passed as a parameter. This
allows one to avoid copying the payload of a union field in a case
where this could be harmful (e.g. the payload includes an allocator).

The implementation of TagPayloadPtrType is not ideal; feel free to
suggest better ways of doing it. I was not even going to make it, but
using `?@TypeOf(&@field(u, @tagName(tag)))` as the return value for
tagPayloadPtr triggered comptime UB-safety errors, and I don't know
enough about how to calculate the type of a derived pointer to
implement TagPayloadPtrType with pointer TypeInfo.