Temporarily fall back to ld linker when native
This commit ensures that when compiling Zig natively on macOS
we use the system linker, `ld`, instead of the LLVM `lld` linker.
Since macOS Big Sur, the OS requires generated binaries (including
dylibs) to be codesigned. If the condition is not met, we receive
an automatic `SIGKILL`.

This commit is hopefully only a temporary solution until either `lld`
fixes this upstream, or self-hosted is in a position to link complex
projects and generate a valid code signature.