stage1: Fix symbol lookup logic
The compiler performed a weird two-pass lookup over the scope hierarchy
when it tried to find a given symbol. Beside being quite inefficient,
doing so is also wrong as it doesn't respect the scope nesting and
introduce lookup rules that defy any sort of common sense.

The newly introduced function performs a single scan over the scope
hierarchy and considers both declarations and variables at the same
time.

The fix is sloppy and inelegant, but that's par for the course as stage1
must die.

Fixes #7140